pyproject.toml
[tool.poetry]
name = "django-inet"
version = "0.5.0.1"
description = "django internet utilities"
readme = "README.md"
repository = "https://github.com/20c/django-inet"
authors = ["20C <code@20c.com>"]
license = "Apache-2.0"

classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Django :: 1.8",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 2",
  "Programming Language :: Python :: 2.7",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.4",
  "Programming Language :: Python :: 3.5",
  "Programming Language :: Python :: 3.6",
  "Topic :: Internet",
  "Topic :: Software Development",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

packages = [{ include = "django_inet", from = "src" }]

[tool.poetry.dependencies]
python = "^3.6.2"

[tool.poetry.dev-dependencies]
# testing
codecov = "^2.0.5"
coverage = "*"
pytest = "^6.0.1"
pytest-django = "^3.8.0"
pytest-cov = "*"
tox = "^3.24.0"
djangorestframework = "^3"

# linting
# bandit = "^1.6.2"
black = { version = ">=20", allow-prereleases = true }
isort = "^5.7.0"
flake8 = "^3.8.4"
mypy = "^0.812"
pre-commit = "^2.13"
pyupgrade = "^2.19.4"

# docs
markdown = "*"
markdown-include = ">=0.5,<1"
mkdocs = "^1"

# ctl
ctl = "^1"
jinja2 = "^2.11.2"
tmpl = "^1"
twine = "^3.3.0"

[tool.poetry.plugins."markdown.extensions"]
pymdgen = "pymdgen.md:Extension"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
